setup-r-dependencies: include LinkingTo packages
I have a GitHub Action that uses `setup-r-dependencies` to install packages based on a `DESCRIPTION` file, and then deploy the app to shinyappsio using {rsconnect}.

The build was failing because my app was using {tidyr} which links to {cpp11}, and {cpp11} was not getting added to the lockfile and was not being installed. But then when the time came to deploy the app, it was failing because of missing {cpp11}. I learned from https://github.com/r-lib/pak/issues/485  how to fix it. I think it makes sense in a github action to make sure that LinkingTo always gets honoured. If you disagree, feel free to close the PR.